requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
schedule>=1.2.0
python-dotenv>=1.0.0
pydantic>=2.0.0
jinja2>=3.1.0